afficherClient, trouverClient, afficherFournisseur, trouverFournisseur views added. Routes implemented for all views
