+++ Wafa/gestionStock/src/main/resources/static/index.html
@@ -24,15 +24,20 @@
 <script type="text/javascript" src="js/app.js"></script>
 <script type="text/javascript" src="js/controllers/ProduitController.js"></script>
 <script type="text/javascript" src="js/controllers/TechnicienController.js"></script>
+<script type="text/javascript" src="js/controllers/ClientController.js"></script>
+<script type="text/javascript" src="js/controllers/FournisseurController.js"></script>
 
 	<!--  directives -->
 <script type="text/javascript" src="js/directives/HeadDirective.js"></script>
 <script type="text/javascript" src="js/directives/tabsDirective.js"></script>
 
 	<!-- Routes -->
+	<script type="text/javascript" src="js/routes/Routes.js"></script>
 	
 	<!-- Services -->
 	<script type="text/javascript" src="js/services/technicienService.js"></script>
+	<script type="text/javascript" src="js/services/clientService.js"></script>
+	<script type="text/javascript" src="js/services/fournisseurService.js"></script>
 	
 </head>
