+++ Wafa/gestionStock/src/main/resources/static/js/services/clientService.js
@@ -1,0 +1,14 @@
+/**
+ * Client services
+ */
+
+app.factory('allClientsService', ['$http', function($http) { 
+  return $http.get('client/tousLesClients') 
+            .success(function(data) { 
+              return data; 
+            }) 
+            .error(function(err) { 
+              return err; 
+            }); 
+}]);
+
