+++ Wafa/gestionStock/src/main/resources/static/js/services/fournisseurService.js
@@ -1,0 +1,14 @@
+/**
+ * Technicien services
+ */
+
+app.factory('allFournisseursService', ['$http', function($http) { 
+  return $http.get('fournisseur/tousLesFournisseurs') 
+            .success(function(data) { 
+              return data; 
+            }) 
+            .error(function(err) { 
+              return err; 
+            }); 
+}]);
+
