+++ Wafa/gestionStock/src/main/resources/static/templates/client/ajouterClientForm.html
