+++ Wafa/gestionStock/src/main/resources/static/templates/shared/entete.html
@@ -18,6 +18,8 @@
 			<li class="dropdown"><a class="dropdown-toggle"
 				data-toggle="dropdown" href="#">Clients<span class="caret"></span></a>
 				<ul class="dropdown-menu">
+					<li><a data-toggle="tab" href="#tousLesClients">Afficher
+							Clients</a></li>
 					<li><a data-toggle="tab" href="#ajouterClient">Ajouter
 							Client</a></li>
 					<li><a data-toggle="tab" href="#trouverClient">Trouver
@@ -26,6 +28,8 @@
 			<li class="dropdown"><a class="dropdown-toggle"
 				data-toggle="dropdown" href="#">Fournisseurs<span class="caret"></span></a>
 				<ul class="dropdown-menu">
+					<li><a data-toggle="tab" href="#tousLesFournisseurs">Afficher
+							Fournisseurs</a></li>
 					<li><a data-toggle="tab" href="#ajouterFournisseur">Ajouter
 							Fournisseur</a></li>
 					<li><a data-toggle="tab" href="#trouverFournisseur">Trouver
