+++ Wafa/gestionStock/src/main/resources/static/templates/shared/home.html
