--- Wafa/gestionStock/src/main/resources/static/toruverFournisseur.html
@@ -1,1 +1,0 @@
-trouver forunisseur form ici
