--- Wafa/gestionStock/src/main/resources/static/tousLesClients.html
@@ -1,2 +1,0 @@
-afficher tous les clients ici : utiliser datatable
-https://datatables.net/examples/api/row_details.html
