--- Wafa/gestionStock/src/main/resources/static/tousLesFournisseurs.html
@@ -1,2 +1,0 @@
-afficher tous les fournisseurs ici : utiliser datatable
-https://datatables.net/examples/api/row_details.html
