--- Wafa/gestionStock/src/main/resources/static/trouverClient.html
@@ -1,1 +1,0 @@
-trouver client
